🎨 Splits constants objects into individual modules
