+++ app/page.tsx
@@ -9,7 +9,7 @@ import { GitHubIcon, LinkedInIcon } from "@/components/social-icons";
 import { Heading } from "@/components/typography/heading";
 import { P } from "@/components/typography/paragraph";
 import { type FrontmatterWithFilename, getAllArticles } from "@/lib/articles";
-import { ROLES } from "@/lib/constants";
+import { ROLES } from "@/lib/constants/roles";
 import { formatDate } from "@/lib/formatDate";
 
 function SocialLink({
