+++ app/uses/page.tsx
@@ -1,7 +1,7 @@
 import { Card } from "@/components/card";
 import { Section } from "@/components/section";
 import { SimpleLayout } from "@/components/simple-layout";
-import { USES } from "@/lib/constants";
+import { USES } from "@/lib/constants/uses";
 
 // https://www.robinwieruch.de/about/ look here for inspo
