+++ components/footer.tsx
@@ -2,7 +2,7 @@ import Link from "next/link";
 
 import { env } from "@/app/env";
 import { ContainerInner, ContainerOuter } from "@/components/container";
-import { PAGES } from "@/lib/constants";
+import { PAGES } from "@/lib/constants/pages";
 
 const NavLink = ({
   href,
