+++ components/resume.tsx
@@ -5,7 +5,7 @@ import Image from "next/image";
 import Link from "next/link";
 
 import { env } from "@/app/env";
-import { type Role } from "@/lib/constants";
+import { type Role } from "@/lib/constants/roles";
 
 const Role = ({ role }: { role: Role }) => {
   const startLabel =
